requirements.txt
openai==2.15.0
python-dotenv==1.2.1
youtube_transcript_api==1.2.3
